requirements.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.6.0
ema-pytorch==0.1.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
gym==0.26.2
gym-notices==0.0.8
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
networkx==3.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
plotly==5.13.0
protobuf==4.21.12
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
sentry-sdk==1.13.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tenacity==8.2.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.13.9
zipp==3.11.0